Rename authentication package to auth
