+++ fabric_rti_mcp/fabric_api_http_client.py
@@ -5,7 +5,7 @@ from typing import Any, cast
 
 import httpx
 
-from fabric_rti_mcp.authentication.auth_context import get_credential
+from fabric_rti_mcp.auth.auth_context import get_credential
 from fabric_rti_mcp.config import GlobalFabricRTIConfig, logger
