+++ fabric_rti_mcp/server.py
@@ -9,7 +9,7 @@ from starlette.requests import Request
 from starlette.responses import JSONResponse
 
 from fabric_rti_mcp import __version__
-from fabric_rti_mcp.authentication.auth_middleware import add_auth_middleware
+from fabric_rti_mcp.auth.auth_middleware import add_auth_middleware
 from fabric_rti_mcp.compat.ms_foundry import SchemaCompatibleMCP
 from fabric_rti_mcp.config import global_config as config
 from fabric_rti_mcp.config import logger
